Issue 31 range rules cover edge cases part 1 (#51)
* Added generation of corner cases for RangeRuleDate

- This is not immutable and if immutability as a principle is accepted
this should be refactored. In spite of being ugly and not very clever,
this is the simplest solution I could think of. Possible solution with
immutability would have to generate new rules and do additional
calculations which would not be very easy to develop nor to read and
reason about.
- Currently this works only with ranges of maximum 4 elements.
- Added integration tests